Title: Suppress bogus GCC 4 uninitialized variable warnings.
Apparently GCC 4 has changed the logic it uses to determine whether C 
variables are being used before they were initialized.  Unfortunately,
it is occasionally emitting warnings for situations that are actually OK.
For example, in any other compiler the following function will not generate
any warnings:

MX_EXPORT mxd_aardvark_get_position( MX_MOTOR *motor )
{
	static const char fname[] = "mxd_aardvark_get_position()";

	MX_AARDVARK *aardvark;
	MX_AARDVARK_MOTOR *aardvark_motor;
	char command[40], response[40];
	int num_items;
	mx_status_type mx_status;

	mx_status = mxd_aardvark_get_pointers( motor,
					&aardvark_motor, &aardvark, fname );

	if ( mx_status.code != MXE_SUCCESS )
		return mx_status;

	sprintf( command, "%dGETPOS", aardvark_motor->axis_number );

	mx_status = mxi_aardvark_command( aardvark, command,
					response, sizeof(response),
					MXD_AARDVARK_DEBUG );

	if ( mx_status.code != MXE_SUCCESS )
		return mx_status;

	num_items = sscanf( response, "%g", &(motor->raw_position.analog) );

	if ( num_items != 1 ) {
		return mx_error( MXE_DEVICE_IO_ERROR, fname,
		"Did not find the position for motor '%s' in the "
		"response '%s' to command '%s'.",
			motor->record->name, response, command );
	}

	return MX_SUCCESSFUL_RESULT;
}

However, every once in a while (somewhere between 1% and 0.1% of the time
I think), GCC 4 emits warning messages of the form:

d_aardvark.c:447: warning: 'aardvark_motor' may be used uninitialized in this function.
d_aardvark.c:446: warning: 'aardvark' may be used uninitialized in this function.

After some examination, it appears that GCC 4 is concerned by the fact that
the first reference to these pointers is in the function call

	mx_status = mxd_aardvark_get_pointers( motor,
					&aardvark_motor, &aardvark, fname );

Now, as it happens, mxd_aardvark_get_pointers()'s whole purpose in life is
to initialize the values of aardvark_motor and aardvark.  Complaining that
aardvark_motor and aardvark were not initialized before the call to
mxd_aardvark_get_pointers() totally misses the point of what is being done
here.  Other compilers assume that if you pass a variable by reference to
a function, that the function may well assign a value to the variable. 
Apparently, GCC 4 sometimes takes the contrary view that there is no 
_guarantee_ that the function initializes the variables.  However, it only
comes to this conclusion every once in a while.  It is fortunate that GCC 4
only does this once in a while, since a crude grep of the MX source code 
indicates that there are around 1969 different places in libMx where this
type of idiom is used.

Anyway, since it only happens occasionally, the simplest fix is to just
to initialize the variables for the rare cases where this warning appears.
Adding code like the following

	aardvark = NULL;
	aardvark_motor = NULL;

is sufficient to shut GCC 4 up.  However, since it is not at all obvious
what distinguishes the cases where GCC 4 emits this warning from those
where it does not, we can expect that there will be more cases in the future
where this bogus warning appears again.  Leaving -Werror out of the GCC
compiler command line is not the answer since most of the time the warnings
GCC emits are useful hints that you may have a subtle bug in your code.

For what it is worth, a variety of other projects on the Net appear to have
been hit by this "bug".  For example, the Linux kernel's Kernel Janitors 
mailing list indicates that they have had to add a variety of redundant
initializers to the Linux kernel itself to shut up GCC 4's complaints
about this.  So I am in good company here.  I guess its just the hazard
of using version x.0 of anything.

Bill Lavender



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@9 42547f41-83fc-0310-9aef-9668e019ab9e