Apparently, a thread in Posix pthreads can be terminated by another thread only by invoking pthread_cancel().  However, the effect of pthread_cancel() on the thread is to immediately take control away from the thread.  Thus, we have a different situation than for Win32 where the main function of the thread can get back control and decide what to do.
For Posix, we instead have available a destructor for mx_current_thread_key
that gets called when the thread is in the process of being destroyed.  If
we modify our model so that we now rely on the asynchronous invocation of
a 'stop request handler' instead of returning a flag when a thread is being
stopped, we can then create a model that both Posix and Win32 can handle.

Thus, we have added a new function mx_thread_set_stop_request_handler()
that configures an optional 'stop request handler' function for the thread.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@21 42547f41-83fc-0310-9aef-9668e019ab9e