Fixed an MCS quick scan bug.  For EPICS synApps 5.1.1, if you send a start command to an EPICS controlled Struck 3801 MCS, there may be an observable delay before the MCS actually starts acquiring measurements.  In some cases, this can result in the quick scan motor starting to move before the MCS starts taking measurements.  This, of course, makes it impossible to correctly determine the position of the motor during the quick scan.
The MX MCS quick scan driver has now been modified to wait until the MCS
says that it is 'busy' acquiring data before starting the quick scan motor.
From preliminary testing, this appears to have fixed the problem.  It is
arguable that this is the way that the MCS quick scan should always have
worked, so this change is probably for the best.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@120 42547f41-83fc-0310-9aef-9668e019ab9e