The network debugging feature can now be turned on or off at run time. If you want to start the debugging at the beginning of a program, add -A to the command line.
If you want to start network debugging on the MX server after it has started,
run the command 'mxput mx_database.network_debug 1'.  Debugging will only
be turned on for network connections made after this flag has been set.

If you want to start network debugging in the 'motor' program after it has
started, find the server you want to start debugging for with 'show server'.
If it has a name like '10id', then run a command like this

    set field 10id.server_flags 0x40000000

If you already had other bits set in the server flags, it would be best 
to preserve them as well.  You can see the preexisting values with the
command

    show field 10id.server_flags



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@585 42547f41-83fc-0310-9aef-9668e019ab9e