mx_pipe() code now compiles on Win32 and seems to run quasi-correctly. The Win32 anonymous pipe seems to not object to being put into non-blocking mode for both the read and write handles.  However, when you attempt to detect the number of bytes in the pipe with PeekNamedPipe(), it seems to return the value 12435000 when the pipe is empty rather than 0.  If you then go on to try to read from the empty pipe with ReadFile, it then returns with the Win32 error code ERROR_NO_DATA 'The pipe is being closed'.  If you ignore the statement 'the pipe is being closed' and take the error code ERROR_NO_DATA at face value, it appears that it is telling you the right thing though.
In spite of all this, the Win32 support seems to work correctly with 
backlash callbacks.  The motor appropriately goes to the backlash position
and then onto the final destination in spite of the torrent of error codes
being returned.  So it may just be a matter of correctly interpreting
the meaning of ERROR_NO_DATA.  However, it would be nice if PeekNamedPipe()
returned the correct number of bytes waiting to be read.



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@770 42547f41-83fc-0310-9aef-9668e019ab9e