Finished a first cut at making the mx_*.h header files safe for C++. This involves adding 'extern "C"' wrappers at appropriate places in the MX headers.  However, this is slightly more complicated than it may initially looks.  Apparently on some systems, it is bad to put includes of system headers such as <stdio.h> inside an 'extern "C"' block.  This means that doing something like this
  #ifdef __cplusplus
  extern "C" {
  #endif

  #include <stdio.h>

  #ifdef __cplusplus
  }
  #endif

can be "bad" on some platforms.

An example of this is in Sun's Solaris documentation.  As of October 2007,
the web page

    http://developers.sun.com/solaris/articles/mixing.html

states the following:

  Warning-

  Do not use this technique for system headers on the Solaris OS.
  The Solaris headers, and all the headers that come with Sun C
  and C++ compilers, are already configured for use with C and C++
  compilers.  You can invalidate declarations in the Solaris
  headers if you specify a linkage.

However, currently the MX header files

  mx_clock.h
  mx_constants.h
  mx_dirent.h
  mx_program_model.h
  mx_record.h
  mx_stdint.h
  mx_util.h

include system header files.  Thus, the MX header files themselves
must be careful not to put system header files inside 'extern "C"'
blocks.  Furthermore, this means that MX header files that include
any of the above MX header files must not include those files inside
an 'extern "C"' block.

This is apparently an important issue.  However, I do not know how
best to document this issue so that it will be remembered later.
Suggestions are welcome.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@960 42547f41-83fc-0310-9aef-9668e019ab9e