Mpatrol-based memory debugging now seems to be working on Windows (tested with Visual Studio .NET 2003 on Windows XP SP3).  Using Mpatrol requires compiling MX with the -MT switch and linking to mpatrolmt.dll.  I also added heapdiff.obj to the linker command line since the Windows version of Mpatrol does not have an equivalent to the /usr/lib/libmptools.a library from the Debian Sid distribution.
On Windows, 'motornet' has only two unfreed allocations, namely the
same two malloc()s in cmd_split_command_line() that were noticed on
Linux.  Once again, the only reason these are unfreed at exit is that
invoking the 'exit' command prevents the code from getting to the end
of the command loop where the two allocations would normally be freed.

As for the MX server, 'msoftsrv -m1' has three unfreed allocations:

1.  mxp_get_total_heap_size() allocates an array of heap handle pointers to
    be used by GetProcessHeaps().  mxp_get_total_heap_size() keeps this 
    pointer around for use by future invocations of mxp_get_total_heap_size(),
    so it is not surprising that it is reported as an unfreed allocation.

2.  According to the Microsoft C runtime source code, gmtime() creates a
    per thread buffer for use as an internal buffer.  The code hangs onto
    this buffer for the sake of future calls to gmtime(), so Mpatrol
    reports it as an unfreed allocation.

3.  A third allocation of 140 bytes is reported to occur in the following
    location:

    0x02CE3858 (140 bytes) {calloc:440:0} [2920|-|-|-]
        0x00687590 _CRT_INIT+305
        0x0068763D _DllMainCRTStartup+93
        0x7C90118A LdrInitializeThunk+36
        0x7C91B195 CsrNewThread+235
        0x7C91B00E RtlUnicodeStringToInteger+381
        0x7C90E457 KiUserApcDispatcher+7

    This appears to somehow be triggered by a call to calloc(), but I'm
    not sure exactly what is going on here.  calloc() does get called when
    an MX network client connects, so its appearance is not completely
    mysterious.  However, since I really do not understand what is going
    on here, I am going to ignore this for now unless the amount of
    unfree memory from this source grows.

One other thing:  Mpatrol does not provide mallinfo(), which means that
the -m1 originally reported 0 for the total amount of allocated memory.
I have created in mx_malloc.c an Mpatrol-specific version of mallinfo()
which only sets the value of mallinfo_struct.uordblks.  It currently gets
this from a call to __mp_stats( &heapinfo_struct ) and does the assignment

    mallinfo_struct.uordblks = heapinfo_struct.atotal;

It appears that this does not do that good a job of emulating uordblks
since the reported amount of memory in use (~ 17150000 bytes) is much
larger than the number of bytes (~ 275000 bytes) reported on Linux.
My suspicion is that Mpatrol is allocating memory in 256 byte blocks.
For the mxserver case, the unfreed allocations is reported to be
216 bytes while heapinfo_struct.atotal reports an increase of 768 bytes.
Since there were 3 unfreed allocations, things would fit if Mpatrol was
allocating memory in 768/3 = 256 byte blocks.  But this is just a
hypothesis.



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1746 42547f41-83fc-0310-9aef-9668e019ab9e