Some motor controllers have the misfeature that just after a move is commanded, they may report that the motor is not moving.  While this may technically be true, it is not useful behavior for higher level code that is waiting for a commanded move to complete.
What we have done is added the concept of a 'busy start interval'.  If
the busy start interval is greater than zero, then after the start of
a motor move, MX will force on the motor->status bit MXSF_MTR_IS_BUSY
for a time interval after the start of the move equal to the busy start 
interval in seconds, regardless of what the motor controller actually
said.  Using this feature requires some experimentation to find the
correct time interval.  Once it is set correctly, this eliminates the
window of time during which the motor reports itself as not moving.

The busy start interval feature is only enabled if the motor field
'busy_start_interval_enabled' is set to a non-zero value.  That lets
most motor drivers ignore the feature.  However, as a side effect,
we now have a way to find out when each motor was last moved.  This
may be useful for administrative purposes.

The motor fields added to support this feature are:

  'busy_start_interval_enabled'  MXFT_BOOL
  'busy_start_interval'          MXFT_DOUBLE in seconds
  'last_start_time'              MXFT_DOUBLE in seconds since some
                                 platform specific 'zero' time.

There are also internal MX_CLOCK tick variables 'busy_start_ticks'
and 'last_start_tick', but these are not visible to external users.



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1829 42547f41-83fc-0310-9aef-9668e019ab9e