The 'daqmx_base' module now appears to work for Red Hat Enterprise Linux 5.
Unfortunately, libMx must be dynamically linked to the National Instruments
DAQmx Base library for the 'daqmx_base' module to work.  Currently, my guess
is that National Instruments is doing something wild and wooly in an
__attribute__((contructor)) function that it expects will happen before
main() is invoked.  If you dlopen() a module, then the constructor is
executed long after main() was invoked.  The symptom when it failed was
that calls to DAQmxBaseCreateTask() would return a NULL TaskHandle.  With
libMx dynamically linked to the National Instruments libraries, this error
no longer occurs.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2007 42547f41-83fc-0310-9aef-9668e019ab9e