Added new flag bits for the 'compumotor_int' driver.  Setting the flag MXF_COMPUMOTOR_DEBUG_SERIAL (0x1000000) causes mxi_compumotor_command() to show high level command and response sequences.  However, there is a call to mx_rs232_getchar() in the response loop which is not turned on by just this flag alone.  This getchar call is used to discard all input until a '*' character is seen, since the '*' character is the prefix to the expected response.
The new MXF_COMPUMOTOR_DEBUG_SERIAL (0x2000000) bit affects the call to
mx_rs232_getchar() mentioned above.  By itself, the bit does nothing.
However, the characters received by this call will be displayed if
one of three things is true:

A.  The MXF_COMPUMOTOR_DEBUG_SERIAL flag is set in the 'compumotor_int'
    record.

B.  The MXF_232_DEBUG_SERIAL bit is set in the 'rs232' record that is
    invoked by the 'compumotor_int' record.

C.  The MXF_232_DEBUG_SERIAL_VERBOSE bit is set in the 'rs232' record
    that is invoked by the 'compumotor_int' record.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2949 42547f41-83fc-0310-9aef-9668e019ab9e