Fixed a bug in mxs_mcs_quick_scan_free_arrays() where we were not correctly discarding the pointer to the old array mcs_quick_scan->motor_position_array when calling this line     (void) mx_free_array( mcs_quick_scan->motor_position_array );
The important issue here is that mx_free() (which is a macro) automatically
overwrites its argument with a NULL pointer.  However, the function
mx_free_array() (which is _NOT_ a macro) does not and cannot overwrite its
argument with a NULL pointer in the context of the function that called it.

Instead, we must explicitly include the line
    mcs_quick_scan->motor_position_array = NULL;

so that later users of the mcs_quick_scan structure are aware that the old
version of mcs_quick_scan->motor_position_array has been freed.

The symptom of this was that mxs_mcs_quick_scan_free_arrays() was trying
to free mcs_quick_scan->motor_position_array, when it had already been 
freed.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3358 42547f41-83fc-0310-9aef-9668e019ab9e