We now have a working 'libtiff' module that saves most information that the end user might like in TIFF tags.  It has been tested using the RDI detector at ALS 12.3.1.
We have not been able to figure out the way to put the start of the image
data at a programmer offset from the start of the image file, since Libtiff
does not expose a method for doing that.  However, the image data appears
to always show up at an offset of 8 bytes from the start of the file, so
that may be good enough.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3560 42547f41-83fc-0310-9aef-9668e019ab9e