The MX 'python' module now uses the contents of 'list_head->program_name' to figure out whether or not it is running in the context of the 'mxmotor' program.
Regrettably, this copy of the MX 'python' module still crashes after one
use.  This appears to be a reference counting bug, since the proximate
cause of the crash appears to be when the module attempts to make use
of an MpMotor.get_position object that has been deleted for some reason.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3661 42547f41-83fc-0310-9aef-9668e019ab9e